Added the victory state (which is just a "next" button for now).

+++ RecoilGame/MachineGun.cs
@@ -53,7 +53,7 @@ namespace RecoilGame
                 Vector2 direction = new Vector2(xNormalized * bulletSpeed, yNormalized * bulletSpeed);
 
                 //Test to see if this will actually create a projectile and how it will work, then we'll add more since we want shotgun to have multiple projectiles
-                new Projectile(player.CenteredX, player.CenteredY, 7, 7, projectileTexture, true, direction, damage, 5, 0.75f, false, true);
+                new Projectile((int)player.CenteredX, (int)player.CenteredY, 7, 7, projectileTexture, true, direction, damage, 5, 0.75f, false, true);
 
                 //Calls playerManager's shooting capability method
                 Game1.playerManager.ShootingCapability();
